+++ release_notes.md
@@ -1,4 +1,5 @@
 ###In Development
+  - [#257](https://github.com/MehdiK/Humanizer/pull/277): Added support for custom enum description attribute property names
 
 [Commits](https://github.com/MehdiK/Humanizer/compare/v1.26.1...master)